Backend/FeedbackAnalysis/requirements.txt
functions-framework==3.*
google-cloud-language
google-api-python-client
google-cloud